✨ Add CursoCreateCommandRequestValidator for validating CursoCreateCommandRequest
The code changes add a new class `CursoCreateCommandRequestValidator` in the `MasterNet.Application.Cursos.CursoCreate` namespace. This validator class is responsible for validating the `CursoCreateCommandRequest` object using FluentValidation. It ensures that the `CursoCreateRequest` property is validated using the `CursoCreateValidator` class. This addition enhances the validation process for creating a new curso.